Link graph file made for visualization
I used the react d3 libraray to define a configuration for my graph and congregate the series data which Metabase uses into data that the graph can read. As of right now there is hardcoded filter of the  Program Name using inter-scope demo data to create the link-graph. The settings are defined in a different file and this file extends another file which helps render the visualization.