Created LinkGraphRender.jsx
This file was copied from LineAreaBarChart.jsx as a template because it has the nessisary methods and checks for the data (series) to transform the data and check if the visualization can be rendered. There may be residual methods and variables needed for LineAreaBarChart which are not needed for the LinkGraph.